+++ simulator/airportcomponent.py
