+++ simulator/gate.py
@@ -7,9 +7,8 @@ import argparse
 from redis import Redis
 
 from restorable import construct_or_restore
-from interfaces import AirportComponent
+from airportcomponent import AirportComponent
 from plane import Plane, PlaneState
-from logger import Logger
 
 REDIS_BROKER = "localhost"
